allow both readme.md and readme.txt to co-exist

+++ deploy.sh
@@ -42,8 +42,12 @@ cd $DEST_DIR
 
 # Transform the readme
 if [ -f readme.md ]; then
-	mv readme.md readme.txt
-	sed -i '' -e 's/^# \(.*\)$/=== \1 ===/' -e 's/ #* ===$/ ===/' -e 's/^## \(.*\)$/== \1 ==/' -e 's/ #* ==$/ ==/' -e 's/^### \(.*\)$/= \1 =/' -e 's/ #* =$/ =/' readme.txt
+	if [ -f readme.txt ]; then
+		echo "Both readme.md and readme.txt found, leaving them alone."
+	else
+		mv readme.md readme.txt
+		sed -i '' -e 's/^# \(.*\)$/=== \1 ===/' -e 's/ #* ===$/ ===/' -e 's/^## \(.*\)$/== \1 ==/' -e 's/ #* ==$/ ==/' -e 's/^### \(.*\)$/= \1 =/' -e 's/ #* =$/ =/' readme.txt
+	fi
 fi
 
 # svn addremove
